+++ src/components/DaoCreate/TokenAllocationInput.tsx
@@ -1,8 +1,8 @@
-import { TokenAllocation } from "../../../hooks/useDeployDAO";
-import { checkAddress } from "../../../hooks/useAddress";
-import { useWeb3 } from "../../../contexts/web3Data";
-import { TextButton } from "../../ui/forms/Button";
-import Input from "../../ui/forms/Input";
+import { TokenAllocation } from "../../hooks/useDeployDAO";
+import { checkAddress } from "../../hooks/useAddress";
+import { useWeb3 } from "../../contexts/web3Data";
+import { TextButton } from "../ui/forms/Button";
+import Input from "../ui/forms/Input";
 
 interface TokenAllocationProps {
   index: number;
