+++ src/components/DaoCreate/TokenDetails.tsx
@@ -1,8 +1,8 @@
 import { useEffect, useCallback, ChangeEvent } from "react";
-import ContentBoxTitle from "../../ui/ContentBoxTitle";
-import { TokenAllocation } from "../../../hooks/useDeployDAO";
-import Input from "../../ui/forms/Input";
-import InputBox from "../../ui/forms/InputBox";
+import ContentBoxTitle from "../ui/ContentBoxTitle";
+import { TokenAllocation } from "../../hooks/useDeployDAO";
+import Input from "../ui/forms/Input";
+import InputBox from "../ui/forms/InputBox";
 import TokenAllocations from "./TokenAllocations";
 
 interface TokenDetailsProps {
